+++ test/MockHttp.Tests/Language/Flow/Response/ServerTimeoutWithNullTimeSpanSpec.cs
@@ -1,0 +1,11 @@
+﻿#nullable enable
+namespace MockHttp.Language.Flow.Response;
+
+public class ServerTimeoutWithNullTimeSpanSpec : ServerTimeoutSpec
+{
+    public ServerTimeoutWithNullTimeSpanSpec()
+        : base(null)
+    {
+    }
+}
+#nullable restore