+++ test/MockHttp.Tests/Language/Flow/Response/ServerTimeoutWithZeroTimeSpanSpec.cs
@@ -1,0 +1,11 @@
+﻿#nullable enable
+namespace MockHttp.Language.Flow.Response;
+
+public class ServerTimeoutWithZeroTimeSpanSpec : ServerTimeoutSpec
+{
+    public ServerTimeoutWithZeroTimeSpanSpec()
+        : base(TimeSpan.Zero)
+    {
+    }
+}
+#nullable restore